+++ CHANGELOG.md
@@ -20,7 +20,8 @@
 - Variable lookup now supports introspection of Swift types. You can now lookup
   values of Swift structures and classes inside a Context.
 
-- If tags can now use prefix and infix operators such as `not`, `and` and `or`.
+- If tags can now use prefix and infix operators such as `not`, `and`, `or` and
+  `==`.
 
     ```html+django
     {% if one or two and not three %}